Auth. System/Queue Separation for TA's/Students

+++ src/main/scala/model/OfficeHoursServer.scala
@@ -6,6 +6,7 @@ import model.database.{Database, DatabaseAPI, TestingDatabase}
 import play.api.libs.json.{JsValue, Json}
 import akka.actor.{Actor, ActorRef, ActorSystem, Props}
 
+//noinspection DuplicatedCode
 class OfficeHoursServer() {
 
   val database: DatabaseAPI = if (Configuration.DEV_MODE) {
@@ -17,6 +18,9 @@ class OfficeHoursServer() {
   var usernameToSocket: Map[String, SocketIOClient] = Map()
   var socketToUsername: Map[SocketIOClient, String] = Map()
 
+  var TA_usernameToSocket: Map[String, SocketIOClient] = Map()
+  var TA_socketToUsername: Map[SocketIOClient, String] = Map()
+
   val config: Configuration = new Configuration {
     setHostname("0.0.0.0")
     setPort(8080)
@@ -27,12 +31,19 @@ class OfficeHoursServer() {
   server.addDisconnectListener(new DisconnectionListener(this))
   server.addEventListener("enter_queue", classOf[String], new EnterQueueListener(this))
   server.addEventListener("ready_for_student", classOf[Nothing], new ReadyForStudentListener(this))
+  server.addEventListener("ta_login", classOf[String], new TALogin(this))
   server.addEventListener("jumbotron", classOf[Nothing], new JumbotronListener(this))
 
   server.start()
 
-  def queueJSON(): String = {
-    val queue: List[StudentInQueue] = database.getQueue
+  def queueJSON_Student(sock: SocketIOClient): String = {
+    val queue: List[StudentInQueue] = database.getStudent_queue(this.socketToUsername.apply(sock))
+    val queueJSON: List[JsValue] = queue.map((entry: StudentInQueue) => entry.asJsValue())
+    Json.stringify(Json.toJson(queueJSON))
+  }
+
+  def queueJSON_TA(): String = {
+    val queue: List[StudentInQueue] = this.database.getQueue
     val queueJSON: List[JsValue] = queue.map((entry: StudentInQueue) => entry.asJsValue())
     Json.stringify(Json.toJson(queueJSON))
   }
@@ -79,29 +90,51 @@ class DisconnectionListener(server: OfficeHoursServer) extends DisconnectListene
   }
 }
 
-
+//noinspection DuplicatedCode
 class EnterQueueListener(server: OfficeHoursServer) extends DataListener[String] {
   override def onData(socket: SocketIOClient, username: String, ackRequest: AckRequest): Unit = {
-    server.database.addStudentToQueue(StudentInQueue(username, System.nanoTime()))
-    server.socketToUsername += (socket -> username)
-    server.usernameToSocket += (username -> socket)
-    server.server.getBroadcastOperations.sendEvent("queue", server.queueJSON())
+    if (!server.TA_socketToUsername.contains(socket) && !server.socketToUsername.contains(socket)) {
+      server.database.addStudentToQueue(StudentInQueue(username, System.nanoTime()))
+      server.socketToUsername += (socket -> username)
+      server.usernameToSocket += (username -> socket)
+      for (i <- server.socketToUsername.keys) i.sendEvent("queue", server.queueJSON_Student(i))
+      for (i <- server.TA_socketToUsername.keys) i.sendEvent("queue", server.queueJSON_TA())
+    }
   }
 }
 
 
+//noinspection DuplicatedCode
 class ReadyForStudentListener(server: OfficeHoursServer) extends DataListener[Nothing] {
   override def onData(socket: SocketIOClient, dirtyMessage: Nothing, ackRequest: AckRequest): Unit = {
     val queue = server.database.getQueue.sortBy(_.timestamp)
-    if (queue.nonEmpty) {
+    if(queue.nonEmpty && server.TA_socketToUsername.contains(socket)){
       val studentToHelp = queue.head
       server.database.removeStudentFromQueue(studentToHelp.username)
       socket.sendEvent("message", "You are now helping " + studentToHelp.username)
       if (server.usernameToSocket.contains(studentToHelp.username)) {
         server.usernameToSocket(studentToHelp.username).sendEvent("message", "A TA is ready to help you")
       }
-      server.server.getBroadcastOperations.sendEvent("queue", server.queueJSON())
+      for (i <- server.socketToUsername.keys) i.sendEvent("queue", server.queueJSON_Student(i))
+      for (i <- server.TA_socketToUsername.keys) i.sendEvent("queue", server.queueJSON_TA())
     }
+  }
+}
+
+class TALogin(server: OfficeHoursServer) extends DataListener[String] {
+  override def onData(socketIOClient: SocketIOClient, t: String, ackRequest: AckRequest): Unit = {
+    val parsed = Json.parse(t)
+    val TA_username = (parsed \ "username").as[String]
+
+    if (!server.TA_usernameToSocket.contains(TA_username) && server.database.checkTACredentials(t)) {
+      server.TA_usernameToSocket += (TA_username -> socketIOClient)
+      server.TA_socketToUsername += (socketIOClient -> TA_username)
+      socketIOClient.sendEvent("queue", server.queueJSON_TA())
+      socketIOClient.sendEvent("release_help_button", "")
+      socketIOClient.sendEvent("login_successful", "")
+    }
+    else if (server.TA_usernameToSocket.contains(TA_username) && server.database.checkTACredentials(t)) socketIOClient.sendEvent("already_logged_in", "")
+    else socketIOClient.sendEvent("login_failed", "")
   }
 }
